+++ commcare/src/main/java/org/motechproject/commcare/tasks/builder/CaseActionBuilder.java
@@ -16,9 +16,9 @@ import java.util.List;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
-import static org.motechproject.tasks.domain.mds.ParameterType.BOOLEAN;
-import static org.motechproject.tasks.domain.mds.ParameterType.MAP;
-import static org.motechproject.tasks.domain.mds.ParameterType.UNICODE;
+import static org.motechproject.tasks.domain.enums.ParameterType.BOOLEAN;
+import static org.motechproject.tasks.domain.enums.ParameterType.MAP;
+import static org.motechproject.tasks.domain.enums.ParameterType.UNICODE;
 
 /**
  * Class responsible for building the Task module actions, related to the case submission.
